chore: add tests; refactor lib
