+++ .github/workflows/ci.yml
@@ -52,3 +52,18 @@ jobs:
         with:
           command: clippy
           args: -- -D warnings
+
+  test:
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v3
+
+      - uses: actions-rs/toolchain@v1
+        with:
+          toolchain: stable
+          profile: minimal
+          override: true
+
+      - name: Run tests
+        run: cargo test --verbose
