+++ src/scanner.rs
